requirements.txt
astroid==1.6.5
ArgumentParser==1.2.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
googleapis-common-protos==1.5.5
grpcio==1.16.0
grpcio-tools==1.16.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
protobuf==3.6.1
pylint==1.9.3
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11